+++ python/twitter_bot/src/classes/types.py
